Move public function to common.py from upgrade_image.py
